+++ db/migrate/20230110062852_add_copyright_info_source_to_site.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddCopyrightInfoSourceToSite < ActiveRecord::Migration[7.0]
+  def change
+    add_column :folio_sites, :copyright_info_source, :string
+  end
+end